pyproject.toml
[project]
name = "multimodal-dedup"
version = "0.1.0"
description = "Multimodal dataset deduplication framework (image/audio/text)"
requires-python = ">=3.10"

# Base: sorter + orchestrator + all pipeline_api cores
dependencies = [
    "numpy>=1.26,<3",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
# Image: CLIP embedding + quality-aware SemDeDup
image = [
    "Pillow>=10.0",
    "torch>=2.1",
    "open-clip-torch>=2.20",
    "scikit-learn>=1.3",
]

# Audio: spectrogram fingerprint + phash/MFCC methods
audio = [
    "librosa>=0.10",
    "soundfile>=0.12",
    "scikit-image>=0.21",
    "matplotlib>=3.7",
    "imagehash>=4.3",
    "Pillow>=10.0",
]

# Text: optional faster dedup methods (default jaccard needs no extras)
text = [
    "datasketch>=1.6",
    "simhash>=2.1",
]

# Install everything
all = [
    "multimodal-dedup[image,audio,text]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipelines", "image", "audio", "text"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
